[data_source] - feature: enhance DataSourceEdition modal functionality
 - Enable managed data source handling directly from DataSourceEdition modal
 - Add ability to set data source integrations and loading state from the modal
 - Check for existence of dataSourceIntegration before proceeding in modal logic

[data_sources/managed] - refactor: remove ConfirmationModal and related functionality

 - Streamline data sources management by utilizing DataSourceEdition modal for confirmations
 - Move REDIRECT_TO_EDIT_PERMISSIONS array to DataSourceEdition to centralize redirection logic
 - Remove unused imports and code related to the now-deprecated ConfirmationModal component
