+++ front/pages/w/[wId]/builder/data-sources/managed.tsx
@@ -1,6 +1,5 @@
 import {
   Avatar,
-  BookOpenIcon,
   Button,
   Chip,
   CloudArrowLeftRightIcon,
@@ -18,7 +17,6 @@ import {
 } from "@dust-tt/sparkle";
 import type {
   ConnectorProvider,
-  DataSourceType,
   EditedByUser,
   ManageDataSourcesLimitsType,
   Result,
@@ -36,23 +34,18 @@ import {
 } from "@dust-tt/types";
 import type { CellContext, ColumnDef } from "@tanstack/react-table";
 import type { InferGetServerSidePropsType } from "next";
-import Link from "next/link";
 import type { NextRouter } from "next/router";
 import { useRouter } from "next/router";
 import { useRef } from "react";
-import { useContext, useEffect, useMemo, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import * as React from "react";
 
-import type {
-  DataSourceIntegration} from "@app/components/data_source/DataSourceEdition";
-import {
-  DataSourceEditionModal
-} from "@app/components/data_source/DataSourceEdition";
+import type { DataSourceIntegration } from "@app/components/data_source/DataSourceEdition";
+import { DataSourceEditionModal } from "@app/components/data_source/DataSourceEdition";
 import ConnectorSyncingChip from "@app/components/data_source/DataSourceSyncChip";
 import { RequestDataSourcesModal } from "@app/components/data_source/RequestDataSourcesModal";
 import { subNavigationBuild } from "@app/components/navigation/config";
 import AppLayout from "@app/components/sparkle/AppLayout";
-import { SendNotificationsContext } from "@app/components/sparkle/Notification";
 import { getDataSourceUsage } from "@app/lib/api/agent_data_sources";
 import config from "@app/lib/api/config";
 import { getDataSources } from "@app/lib/api/data_sources";
@@ -61,7 +54,6 @@ import { withDefaultUserAuthRequirements } from "@app/lib/iam/session";
 import { useAdmins } from "@app/lib/swr";
 import { classNames, timeAgoFrom } from "@app/lib/utils";
 import logger from "@app/logger/logger";
-import type { PostManagedDataSourceRequestBody } from "@app/pages/api/w/[wId]/data_sources/managed";
 
 const { GA_TRACKING_ID = "" } = process.env;
 
@@ -89,6 +81,11 @@ type GetTableRowParams = {
   readOnly: boolean;
 };
 
+type ShowIntegration = {
+  show: boolean;
+  dataSourceIntegration: DataSourceIntegration | null;
+};
+
 type HandleConnectionClickParams = {
   integration: DataSourceIntegration;
   isAdmin: boolean;
@@ -97,7 +94,7 @@ type HandleConnectionClickParams = {
   owner: WorkspaceType;
   limits: ManageDataSourcesLimitsType;
   setShowUpgradePopup: (show: boolean) => void;
-  setShowConfirmConnection: (integration: DataSourceIntegration | null) => void;
+  setShowEditionDataSourceModalOpen: (showIntegration: ShowIntegration) => void;
   setShowPreviewPopupForProvider: (providerPreview: {
     show: boolean;
     connector: ConnectorProvider | null;
@@ -113,14 +110,6 @@ type ManagedSourceType = {
   editedByUser?: EditedByUser | null;
   usage: number | null;
 };
-
-const REDIRECT_TO_EDIT_PERMISSIONS = [
-  "confluence",
-  "google_drive",
-  "microsoft",
-  "slack",
-  "intercom",
-];
 
 export async function setupConnection({
   dustClientFacingUrl,
@@ -323,132 +312,6 @@ export const getServerSideProps = withDefaultUserAuthRequirements<{
   };
 });
 
-function ConfirmationModal({
-  dataSource,
-  show,
-  onClose,
-  onConfirm,
-}: {
-  dataSource: DataSourceIntegration;
-  show: boolean;
-  onClose: () => void;
-  onConfirm: () => void;
-}) {
-  const [isLoading, setIsLoading] = useState(false);
-
-  return (
-    <Modal
-      isOpen={show}
-      title={`Connect ${dataSource.name}`}
-      onClose={onClose}
-      hasChanged={false}
-      variant="side-sm"
-    >
-      <div className="pt-8">
-        <Page.Vertical gap="lg" align="stretch">
-          <div className="flex flex-col gap-y-2">
-            <div className="grow text-sm font-medium text-element-800">
-              Important
-            </div>
-            <div className="text-sm font-normal text-element-700">
-              Resources shared with Dust will be made available to the entire
-              workspace{" "}
-              <span className="font-medium">
-                irrespective of their granular permissions
-              </span>{" "}
-              on {dataSource.name}.
-            </div>
-          </div>
-
-          {dataSource.limitations && (
-            <div className="flex flex-col gap-y-2">
-              <div className="grow text-sm font-medium text-element-800">
-                Limitations
-              </div>
-              <div className="text-sm font-normal text-element-700">
-                {dataSource.limitations}
-              </div>
-            </div>
-          )}
-
-          {dataSource.connectorProvider === "google_drive" && (
-            <>
-              <div className="flex flex-col gap-y-2">
-                <div className="grow text-sm font-medium text-element-800">
-                  Disclosure
-                </div>
-                <div className="text-sm font-normal text-element-700">
-                  Dust's use of information received from the Google APIs will
-                  adhere to{" "}
-                  <Link
-                    className="s-text-action-500"
-                    href="https://developers.google.com/terms/api-services-user-data-policy#additional_requirements_for_specific_api_scopes"
-                  >
-                    Google API Services User Data Policy
-                  </Link>
-                  , including the Limited Use requirements.
-                </div>
-              </div>
-
-              <div className="flex flex-col gap-y-2">
-                <div className="grow text-sm font-medium text-element-800">
-                  Notice on data processing
-                </div>
-                <div className="text-sm font-normal text-element-700">
-                  By connecting Google Drive, you acknowledge and agree that
-                  within your Google Drive, the data contained in the files and
-                  folders that you choose to synchronize with Dust will be
-                  transmitted to third-party entities, including but not limited
-                  to Artificial Intelligence (AI) model providers, for the
-                  purpose of processing and analysis. This process is an
-                  integral part of the functionality of our service and is
-                  subject to the terms outlined in our Privacy Policy and Terms
-                  of Service.
-                </div>
-              </div>
-            </>
-          )}
-
-          <div className="flex justify-center pt-2">
-            <Button.List isWrapping={true}>
-              <Button
-                variant="primary"
-                size="md"
-                icon={CloudArrowLeftRightIcon}
-                onClick={() => {
-                  setIsLoading(true);
-                  onConfirm();
-                }}
-                disabled={isLoading}
-                label={
-                  isLoading
-                    ? "Connecting..."
-                    : dataSource.connectorProvider === "google_drive"
-                      ? "Acknowledge and Connect"
-                      : "Connect"
-                }
-              />
-              {dataSource.guideLink && (
-                <Button
-                  label="Read our guide"
-                  size="md"
-                  variant="tertiary"
-                  icon={BookOpenIcon}
-                  onClick={() => {
-                    if (dataSource.guideLink) {
-                      window.open(dataSource.guideLink, "_blank");
-                    }
-                  }}
-                />
-              )}
-            </Button.List>
-          </div>
-        </Page.Vertical>
-      </div>
-    </Modal>
-  );
-}
-
 export default function DataSourcesView({
   owner,
   subscription,
@@ -460,7 +323,6 @@ export default function DataSourcesView({
   dustClientFacingUrl,
   user,
 }: InferGetServerSidePropsType<typeof getServerSideProps>) {
-  const sendNotification = useContext(SendNotificationsContext);
   const [dataSourceIntegrations, setDataSourceIntegrations] =
     useState(integrations);
   const [isLoadingByProvider, setIsLoadingByProvider] = useState<
@@ -474,13 +336,9 @@ export default function DataSourcesView({
       show: false,
       connector: null,
     });
-  const [showConfirmConnection, setShowConfirmConnection] =
-    useState<DataSourceIntegration | null>(null);
+
   const [showEditionDataSourceModalOpen, setShowEditionDataSourceModalOpen] =
-    useState<{
-      show: boolean;
-      dataSourceIntegration: DataSourceIntegration | null;
-    }>({
+    useState<ShowIntegration>({
       show: false,
       dataSourceIntegration: null,
     });
@@ -489,83 +347,6 @@ export default function DataSourcesView({
 
   const { admins, isAdminsLoading } = useAdmins(owner);
   const planConnectionsLimits = plan.limits.connections;
-  const handleEnableManagedDataSource = async (
-    provider: ConnectorProvider,
-    suffix: string | null
-  ) => {
-    try {
-      const connectionIdRes = await setupConnection({
-        dustClientFacingUrl,
-        owner,
-        provider,
-      });
-      if (connectionIdRes.isErr()) {
-        throw connectionIdRes.error;
-      }
-
-      setShowConfirmConnection(null);
-      setIsLoadingByProvider((prev) => ({ ...prev, [provider]: true }));
-
-      const res = await fetch(
-        suffix
-          ? `/api/w/${
-              owner.sId
-            }/data_sources/managed?suffix=${encodeURIComponent(suffix)}`
-          : `/api/w/${owner.sId}/data_sources/managed`,
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({
-            provider,
-            connectionId: connectionIdRes.value,
-            name: undefined,
-            configuration: null,
-          } satisfies PostManagedDataSourceRequestBody),
-        }
-      );
-
-      if (res.ok) {
-        const createdManagedDataSource: {
-          dataSource: DataSourceType;
-          connector: ConnectorType;
-        } = await res.json();
-        setDataSourceIntegrations((prev) =>
-          prev.map((ds) => {
-            return ds.connector === null && ds.connectorProvider == provider
-              ? {
-                  ...ds,
-                  connector: createdManagedDataSource.connector,
-                  setupWithSuffix: null,
-                  dataSourceName: createdManagedDataSource.dataSource.name,
-                }
-              : ds;
-          })
-        );
-        if (REDIRECT_TO_EDIT_PERMISSIONS.includes(provider)) {
-          void router.push(
-            `/w/${owner.sId}/builder/data-sources/${createdManagedDataSource.dataSource.name}?edit_permissions=true`
-          );
-        }
-      } else {
-        const responseText = await res.text();
-        sendNotification({
-          type: "error",
-          title: `Failed to enable connection (${provider})`,
-          description: `Got: ${responseText}`,
-        });
-      }
-    } catch (e) {
-      setShowConfirmConnection(null);
-      sendNotification({
-        type: "error",
-        title: `Failed to enable connection (${provider})`,
-      });
-    } finally {
-      setIsLoadingByProvider((prev) => ({ ...prev, [provider]: false }));
-    }
-  };
 
   useEffect(() => {
     setDataSourceIntegrations(dataSourceIntegrations);
@@ -675,19 +456,6 @@ export default function DataSourcesView({
           </div>
         </Modal>
       )}
-      {showConfirmConnection && (
-        <ConfirmationModal
-          dataSource={showConfirmConnection}
-          show={true}
-          onClose={() => setShowConfirmConnection(null)}
-          onConfirm={async () => {
-            await handleEnableManagedDataSource(
-              showConfirmConnection.connectorProvider as ConnectorProvider,
-              showConfirmConnection.setupWithSuffix
-            );
-          }}
-        />
-      )}
       <Page.Vertical gap="xl" align="stretch">
         <Page.Header
           title="Connections"
@@ -761,7 +529,7 @@ export default function DataSourcesView({
                         owner,
                         limits: planConnectionsLimits,
                         setShowUpgradePopup,
-                        setShowConfirmConnection,
+                        setShowEditionDataSourceModalOpen,
                         setShowPreviewPopupForProvider,
                       });
                     }}
@@ -811,6 +579,8 @@ export default function DataSourcesView({
           dustClientFacingUrl={dustClientFacingUrl}
           user={user}
           setIsRequestDataSourceModalOpen={setIsRequestDataSourceModalOpen}
+          setDataSourceIntegrations={setDataSourceIntegrations}
+          setIsLoadingByProvider={setIsLoadingByProvider}
         />
       </Page.Vertical>
       {showUpgradePopup && (
@@ -1019,7 +789,7 @@ function handleConnectionClick({
   isAdmin,
   isLoadingByProvider,
   setShowUpgradePopup,
-  setShowConfirmConnection,
+  setShowEditionDataSourceModalOpen,
   setShowPreviewPopupForProvider,
   router,
   owner,
@@ -1040,7 +810,10 @@ function handleConnectionClick({
       setShowUpgradePopup(true);
     } else {
       if (isBuilt) {
-        setShowConfirmConnection(integration);
+        setShowEditionDataSourceModalOpen({
+          show: true,
+          dataSourceIntegration: integration,
+        });
       } else {
         setShowPreviewPopupForProvider({
           show: true,
